Split package workspace/env to pkg/constants and pkg/provision
Split common env var package to merge it into existing pkg/ structure:

* Constants (env var names) to pkg/constants/env.go
* Function providing common env vars to pkg/provision/workspace to sit
  alongside alongside where it's used.
